Re-adding provider will adopt old hosts
When an Infrastructure Provider is deleted, the hosts associated with that
Provider are left behind with a nil ems_id.  If the same Infrastructure Provider
is re-added, it is supposed to pick up the old orphaned hosts.

f4aa372a570c934f99a5d789aaffafb8ac84c6c9 introduced a bug where re-adding a
previously deleted Provider would fail to pick up orphaned hosts.  It resulted
in the following situation:

1. Add Provider for the first time

   Hosts
   1: name:"hostname.domain.com", ems_id:1

2. Delete the provider

   Hosts
   1: name:"hostname.domain.com", :ems_id:nil

3. Re-add the provider

   Hosts
   1: name:"hostname.domain.com", ems_id:nil
   2: name:"hostname.domain.com - 1", ems_id:2

This patch changes the third step to:

3. Re-add the provider

   Hosts
   1: name:"hostname.domain.com", ems_id:2

https://bugzilla.redhat.com/show_bug.cgi?id=1283195