Add migration to fix host_storages replication issue
After we run the migration added in commit 04f2fa211b8ebeb0557ca3712af846c9b556869f
the table host_storages will be assigned a primary key column and existing
rows will be given a value within the current region's range.

This causes an issue when run on a replication target as the id should
match the one on the replication source region not the target one.

This fix removes the entries in the host_storages table on the target
region and resets the replication triggers on the source regions.

Upstream PR: https://github.com/ManageIQ/manageiq/pull/5760
Clean cherry-pick from upstream master
https://bugzilla.redhat.com/show_bug.cgi?id=1289200