Use DRb's builtin caching mechanism
https://bugzilla.redhat.com/show_bug.cgi?id=1297499

Previously we have been using our own caching mechanism to
cache objects sent as references to the Automate Methods. We have
gotten DRb recycled object errors, since some of the objects which
were sent as references were not being cached.
DRb has a builtin caching mechanism based on TimerIdConv, which
we are starting to use in this PR.

Conflicts:
	spec/lib/miq_automation_engine/miq_ae_service_spec.rb