Patch RSpec 3 'expect' syntax
Downstream 5.5.z will not be getting all of the upstream RSpec 3
updates. This isn't a huge deal as most of the expect syntax of RSpec 3
in the upstream project is compatible with RSpec 2.14 in 5.5.z. However,
the one-liner is_expected syntax introduced in RSpec 3.0.0 is fairly
common and having to change it for every backported commit would be
annoying. Thus, we add it here ourselves.