Fix searching for Hosts by vmm_vendor
Rbac will try to optimize a search in SQL wherever
possible. Searching for hosts by vmm_vendor is complicated by the fact
that this data is wrapped in some additional behavior that changes the
case when getting/setting. Because it filters in both SQL and ruby, it
won't ever meet both criteria because the field will be presented in
different cases.

This is a backport of 76e9afe3. This revision was later reverted
upstream in favor of renaming the method that wrapped the field
vmm_vendor to something else; indeed, this was a necessary change
because of the changes to virtual columns brought about by the Rails 5
upgrade. Since this stream will not support Rails 5, the above should
suffice.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1291880

(cherry picked from commit 76e9afe3e0bc1f168067e11afba9d35ae61c1af8)